requirements.txt
Flask==0.10.1
numpy==1.14.3
Pillow==5.1.0
matplotlib==2.2.2
scikit-learn==0.19.1
scipy==1.1.0
requests==2.18.4